==> 01_MIMIC-IV_Microbiology_Processing.ipynb <==
import os
import glob
import pandas as pd
from tqdm import tqdm

# Define paths to the folders
hosp_path = "../mimic-iv-3.1/hosp/*.csv.gz"
icu_path = "../mimic-iv-3.1/icu/*.csv.gz"

# Get list of all csv.gz files
hosp_files = glob.glob(hosp_path)
icu_files = glob.glob(icu_path)

# Combine all file paths
all_files = hosp_files + icu_files
for file in all_files:
    print(file)


pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

# Microbiology Events

### Load and Filter

file_name = '../mimic-iv-3.1/hosp/microbiologyevents.csv.gz'
microbiology_df = pd.read_csv(file_name, compression="gzip")
microbiology_df.head()

# Define the list of target bacteria
target_bacteria = ['ESCHERICHIA COLI', 'ENTEROCOCCUS FAECALIS', 'KLEBSIELLA PNEUMONIAE', 'PSEUDOMONAS AERUGINOSA']

# Filter the dataframe for rows where 'org_name' is in the target list
filtered_microbiology_df = microbiology_df[microbiology_df['org_name'].isin(target_bacteria)]
filtered_microbiology_df.shape

filtered_microbiology_df.head()

# Ensure charttime is in datetime format
filtered_microbiology_df['charttime'] = pd.to_datetime(filtered_microbiology_df['charttime'])
filtered_microbiology_df['chartdate'] = pd.to_datetime(filtered_microbiology_df['chartdate'])

# Fill NaT values in charttime with chartdate
filtered_microbiology_df['charttime'].fillna(filtered_microbiology_df['chartdate'], inplace=True)

# Find the earliest charttime per (subject_id, hadm_id)
earliest_charttime_idx = filtered_microbiology_df.groupby(['subject_id', 'hadm_id','org_name','ab_name'], dropna = False)['charttime'].idxmin()

# Filter down to only those rows
earliest_microbiology_df = filtered_microbiology_df.loc[earliest_charttime_idx]

earliest_microbiology_df.shape

earliest_microbiology_df = earliest_microbiology_df[['subject_id', 'hadm_id','org_name','ab_name','charttime','interpretation']]
earliest_microbiology_df.head(10)

### EDA

# Group by antibiotic name and calculate required metrics
antibiotic_summary_df = (
    earliest_microbiology_df
    .groupby(['org_name','ab_name'])
    .agg(
        row_count=('ab_name', 'count'),
        interpretation_S=('interpretation', lambda x: (x == 'S').sum()),
        interpretation_I=('interpretation', lambda x: (x == 'I').sum()),
        interpretation_R=('interpretation', lambda x: (x == 'R').sum()),
        interpretation_null=('interpretation', lambda x: x.isna().sum()),
    )
    .reset_index()
)

# Calculate the percentage of resistant cases (R) over total row count per antibody
antibiotic_summary_df['percent_resistant'] = (
    (antibiotic_summary_df['interpretation_R'] / antibiotic_summary_df['row_count']) * 100
).round(2)

# Sort by row count in descending order
antibiotic_summary_df.sort_values(by='row_count', ascending=False, inplace = True)


# target_bacteria = ['ESCHERICHIA COLI', 'ENTEROCOCCUS FAECALIS', 'KLEBSIELLA PNEUMONIAE', 'PSEUDOMONAS AERUGINOSA']
antibiotic_summary_df[antibiotic_summary_df['org_name']=='ESCHERICHIA COLI']

# target_bacteria = ['ESCHERICHIA COLI', 'ENTEROCOCCUS FAECALIS', 'KLEBSIELLA PNEUMONIAE', 'PSEUDOMONAS AERUGINOSA']
antibiotic_summary_df[antibiotic_summary_df['org_name']=='ENTEROCOCCUS FAECALIS']

# target_bacteria = ['ESCHERICHIA COLI', 'ENTEROCOCCUS FAECALIS', 'KLEBSIELLA PNEUMONIAE', 'PSEUDOMONAS AERUGINOSA']
antibiotic_summary_df[antibiotic_summary_df['org_name']=='KLEBSIELLA PNEUMONIAE']

# target_bacteria = ['ESCHERICHIA COLI', 'ENTEROCOCCUS FAECALIS', 'KLEBSIELLA PNEUMONIAE', 'PSEUDOMONAS AERUGINOSA']
antibiotic_summary_df[antibiotic_summary_df['org_name']=='PSEUDOMONAS AERUGINOSA']

### AB Filter

# Filter to most common antibiotics
list_1 = ['AMPICILLIN', 'AMPICILLIN/SULBACTAM', 'CEFAZOLIN', 'CEFEPIME', 'CEFTAZIDIME', 'CEFTRIAXONE', 'CIPROFLOXACIN', 'GENTAMICIN', 'MEROPENEM', 'NITROFURANTOIN', 'PIPERACILLIN/TAZO', 'TOBRAMYCIN', 'TRIMETHOPRIM/SULFA']
list_2 = ['AMPICILLIN', 'DAPTOMYCIN', 'LINEZOLID', 'NITROFURANTOIN', 'PENICILLIN G', 'TETRACYCLINE', 'VANCOMYCIN']
list_3 = ['AMPICILLIN/SULBACTAM', 'CEFAZOLIN', 'CEFEPIME', 'CEFTAZIDIME', 'CEFTRIAXONE', 'CIPROFLOXACIN', 'GENTAMICIN', 'MEROPENEM', 'NITROFURANTOIN', 'PIPERACILLIN/TAZO', 'TOBRAMYCIN', 'TRIMETHOPRIM/SULFA']
list_4 = ['CEFEPIME', 'CEFTAZIDIME', 'CIPROFLOXACIN', 'GENTAMICIN', 'MEROPENEM', 'PIPERACILLIN/TAZO', 'TOBRAMYCIN']

ab_list = list(set(list_1 + list_2 + list_3 + list_4))
len(ab_list)

final_microbiology_df = earliest_microbiology_df[earliest_microbiology_df['ab_name'].isin(ab_list)]
final_microbiology_df['susceptible_flag'] = final_microbiology_df['interpretation'].map({'S': 1}).fillna(0)
final_microbiology_df.shape

final_microbiology_df.head(10)

# Build Driver file (list of of subject and hadms)
subject_hadm_driver = earliest_microbiology_df[['subject_id', 'hadm_id']].drop_duplicates()
subject_hadm_time_driver = earliest_microbiology_df[['subject_id', 'hadm_id','charttime']].drop_duplicates()

print(subject_hadm_driver.shape)
print(subject_hadm_time_driver.shape)

# Export
final_microbiology_df.to_pickle("../mimic-iv-3.1/final_microbiology_df.pkl")
subject_hadm_driver.to_pickle("../mimic-iv-3.1/subject_hadm_driver.pkl")
subject_hadm_time_driver.to_pickle("../mimic-iv-3.1/subject_hadm_time_driver.pkl")
