==> 02_MIMIC-IV_Labevents_Processing_2025_03_16.ipynb <==
import os
import glob
import pandas as pd
from tqdm import tqdm
import pickle

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

# Define paths to the folders
hosp_path = "../mimic-iv-3.1/hosp/*.csv.gz"
icu_path = "../mimic-iv-3.1/icu/*.csv.gz"

# Get list of all csv.gz files
hosp_files = glob.glob(hosp_path)
icu_files = glob.glob(icu_path)

# Combine all file paths
all_files = hosp_files + icu_files
for file in all_files:
    print(file)


### Import

# Import
targets = pd.read_pickle("../mimic-iv-3.1/final_microbiology_df.pkl")
# subject_hadm_driver = pd.read_pickle("../mimic-iv-3.1/subject_hadm_driver.pkl")
# subject_hadm_time_driver = pd.read_pickle("../mimic-iv-3.1/subject_hadm_time_driver.pkl")

# Labevents

### Read File and Filter

file_name = '../mimic-iv-3.1/hosp/labevents.csv.gz'
labevents_df = pd.read_csv(file_name, compression="gzip")
labevents_df

# Set index for faster lookups
labevents_df = labevents_df.set_index(['subject_id', 'hadm_id'])

# Filter only valid (subject_id, hadm_id) pairs
labevents_df = labevents_df.loc[
    labevents_df.index.intersection(subject_hadm_driver.set_index(['subject_id', 'hadm_id']).index)
].reset_index()

labevents_df.shape

# Export
labevents_df.to_pickle("../mimic-iv-3.1/labevents_df_4b.pkl")

### Merge with description file

# Import
labevents_df = pd.read_pickle("../mimic-iv-3.1/labevents_df_4b.pkl")
labevents_df.shape

file_name = '../mimic-iv-3.1/hosp/d_labitems.csv.gz'
labitems_description = pd.read_csv(file_name, compression="gzip")
labitems_description.head()

# Set index for faster lookups
labevents_df = labevents_df.set_index('itemid')

# Set index for labitems_description and select only necessary columns
labitems_description = labitems_description.set_index('itemid')[['label', 'fluid', 'category']]

# Perform an index-based join to bring in additional columns
labevents_df = labevents_df.join(labitems_description, how='inner').reset_index()

# Check shape
labevents_df.shape

labevents_df.head()

# Export the reduced labevents file for faster loading and efficiency
labevents_df.to_pickle("../mimic-iv-3.1/labevents_df_4b_wd.pkl")
labevents_df.to_csv("../mimic-iv-3.1/labevents_df_4b_wd.csv.gz", compression="gzip", index=False)

### Find Total Occurences to get most popular lab tests

# Reduce to a single test
# Find the earliest charttime per (subject_id, hadm_id)
earliest_charttime_idx = labevents_df.groupby(['subject_id', 'hadm_id','label','fluid'], dropna = False)['charttime'].idxmin()

# Filter down to only those rows
earliest_labevents_df = labevents_df.loc[earliest_charttime_idx]

earliest_labevents_df.shape

# Count distinct subject_id and hadm_id per label, fluid, and category
distinct_counts = earliest_labevents_df.groupby(['label', 'fluid', 'category']).agg(
    distinct_subjects=('subject_id', 'nunique'),
    distinct_hadm_ids=('hadm_id', 'nunique'),
    total_occurrences=('subject_id', 'count')  # Optional: total number of occurrences
).reset_index()

# Sort by distinct hadm_id count (or distinct subject_id count)
distinct_counts.sort_values(by='distinct_subjects', ascending=False, inplace=True)
distinct_counts.head(100)

# Select the top 100 most frequent "Fluid - Label" pairs
top_100_labs = distinct_counts.head(100)[["label", "fluid"]].values.tolist()
top_100_labs

# Export to a pickle file
with open("../mimic-iv-3.1/top_100_labs.pkl", "wb") as f:
    pickle.dump(top_100_labs, f)

# For Import
# with open("../mimic-iv-3.1/top_100_labs.pkl", "rb") as f:
#     top_100_labs = pickle.load(f)
